+++ android/src/main/kotlin/io/adrop/adrop_ads/AdropAdManager.kt
@@ -12,8 +12,8 @@ class AdropAdManager {
 
     private val ads: MutableMap<String, AdropAd?> = mutableMapOf()
 
-    fun load(context: Context, adType: AdType, unitId: String, requestId: String, messenger: BinaryMessenger) {
-        val ad = getAd(adType, requestId) ?: createAd(context, adType, unitId, requestId, messenger)
+    fun load(context: Context, adType: AdType, unitId: String, requestId: String, useCustomClick: Boolean = false, messenger: BinaryMessenger) {
+        val ad = getAd(adType, requestId) ?: createAd(context, adType, unitId, requestId, useCustomClick, messenger)
         val key = keyOf(adType, requestId)
         ads[key]?:let {
             ads[key] = ad
@@ -45,12 +45,12 @@ class AdropAdManager {
         }
     }
 
-    private fun createAd(context: Context, adType: AdType, unitId: String, requestId: String, messenger: BinaryMessenger): AdropAd? {
+    private fun createAd(context: Context, adType: AdType, unitId: String, requestId: String, useCustomClick: Boolean, messenger: BinaryMessenger): AdropAd? {
         return when (adType) {
             AdType.Interstitial -> FlutterAdropInterstitialAd(context, unitId, requestId, messenger)
             AdType.Rewarded -> FlutterAdropRewardedAd(context, unitId, requestId, messenger)
             AdType.Popup -> FlutterAdropPopupAd(context, unitId, requestId, messenger)
-            AdType.Native -> FlutterAdropNativeAd(context, unitId, requestId, messenger)
+            AdType.Native -> FlutterAdropNativeAd(context, unitId, requestId, useCustomClick, messenger)
             AdType.Undefined -> null
         }
     }