+++ lib/src/native/adrop_native_ad_view.dart
@@ -1,10 +1,12 @@
 import 'package:adrop_ads_flutter/adrop_ads_flutter.dart';
 import 'package:adrop_ads_flutter/src/bridge/adrop_channel.dart';
+import 'package:adrop_ads_flutter/src/bridge/adrop_method.dart';
 import 'package:adrop_ads_flutter/src/model/call_create_ad.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 
+import '../adrop_ad.dart';
 import 'adrop_native_ad_provider.dart';
 
 /// AdropNativeAdView class responsible for displaying native ads to the user.
@@ -21,13 +23,30 @@ class AdropNativeAdView extends StatelessWidget {
   Widget build(BuildContext context) {
     if (ad == null) return Container();
 
+    var useCustomClick = ad?.useCustomClick ?? false;
+    var platformView = Positioned.fill(
+        child: IgnorePointer(ignoring: useCustomClick, child: _platformView()));
+    var children = useCustomClick
+        ? [
+            platformView,
+            Listener(
+              behavior: HitTestBehavior.deferToChild,
+              onPointerDown: (_) {
+                const MethodChannel(AdropChannel.invokeChannel)
+                    .invokeMethod(AdropMethod.performClick, {
+                  "adType": AdType.native.name,
+                  "requestId": ad?.requestId,
+                });
+              },
+              child: child,
+            )
+          ]
+        : [child, platformView];
+
     return AdropNativeAdProvider(
       ad: ad,
       child: Stack(
-        children: [
-          child,
-          Positioned.fill(child: _platformView()),
-        ],
+        children: children,
       ),
     );
   }